feat(approval): refactor account approval flow and add reject functionality (#284)
### TL;DR
Implemented approval and rejection functionality for account requests in the admin dashboard.

### What changed?
- Created a new `ActionRequestButton` component that handles both approval and rejection actions
- Updated the approval columns to show an eye icon for viewing request details
- Consolidated the approval/rejection logic into a single reusable component
- Added success and error toast notifications for user feedback
- Implemented loading states during request processing

### How to test?
1. Navigate to the admin dashboard's approval request section
2. Click the eye icon on any pending account request
3. Review the account details in the modal
4. Test both "Approve" and "Reject" buttons
5. Verify that appropriate success messages appear
6. Confirm that the request status updates accordingly
7. Check that loading states display during processing

### Why make this change?
To streamline the account approval workflow by:
- Providing a more intuitive user interface for reviewing requests
- Reducing code duplication by combining approval/rejection logic
- Improving user feedback through loading states and notifications
- Ensuring data consistency when processing requests
