--- src/app/(dashboard)/admin/approval-request/accounts/approve-request-button.tsx
@@ -1,133 +1,0 @@
-import { useApprovalRequestContext } from '@/app/(dashboard)/admin/approval-request/accounts/approval-request-provider'
-import { useToast } from '@/components/ui/use-toast'
-import { createBrowserClient } from '@/utils/supabase'
-import {
-  useInsertMutation,
-  useUpdateMutation,
-} from '@supabase-cache-helpers/postgrest-react-query'
-import { FC, ReactNode, useEffect } from 'react'
-
-interface ApproveRequestButtonProps {
-  children: ReactNode
-}
-
-const ApproveRequestButton: FC<ApproveRequestButtonProps> = ({ children }) => {
-  const supabase = createBrowserClient()
-  const { toast } = useToast()
-  const { selectedData, setIsModalOpen, setIsLoading } =
-    useApprovalRequestContext()
-
-  const {
-    mutateAsync: insertAccount,
-    isError: isInsertAccountError,
-    isPending: isInsertingAccount,
-  } = useInsertMutation(
-    // @ts-expect-error
-    supabase.from('accounts'),
-    ['id'],
-    null,
-    {
-      onSuccess: () => {
-        toast({
-          title: 'Request approved',
-          description: 'The request has been approved successfully',
-        })
-        setIsModalOpen(false)
-      },
-      onError: () => {
-        toast({
-          title: 'Error',
-          description: 'An error occurred while approving the request',
-        })
-      },
-    },
-  )
-
-  const {
-    mutateAsync: updatePendingAccount,
-    isPending: isUpdatingPendingAccount,
-  } = useUpdateMutation(
-    // @ts-expect-error
-    supabase.from('pending_accounts'),
-    ['id'],
-    null,
-    {
-      onSuccess: () => {
-        toast({
-          title: 'Request approved',
-          description: 'The request has been approved successfully',
-        })
-      },
-      onError: () => {
-        toast({
-          title: 'Error',
-          description: 'An error occurred while approving the request',
-        })
-      },
-    },
-  )
-
-  const handleClick = async () => {
-    if (!selectedData?.id) throw new Error('ID is required')
-
-    // Insert account
-    await insertAccount([
-      {
-        company_name: selectedData.company_name,
-        is_active: selectedData.is_active,
-        agent_id: (selectedData as any).agent.user_id,
-        company_address: selectedData.company_address,
-        nature_of_business: selectedData.nature_of_business,
-        hmo_provider_id: (selectedData as any).hmo_provider.id,
-        previous_hmo_provider_id: (selectedData as any).previous_hmo_provider
-          .id,
-        current_hmo_provider_id: (selectedData as any).current_hmo_provider.id,
-        account_type_id: (selectedData as any).account_type.id,
-        total_utilization: selectedData.total_utilization,
-        total_premium_paid: selectedData.total_premium_paid,
-        signatory_designation: selectedData.signatory_designation,
-        contact_person: selectedData.contact_person,
-        contact_number: selectedData.contact_number,
-        principal_plan_type_id: (selectedData as any).principal_plan_type.id,
-        dependent_plan_type_id: (selectedData as any).dependent_plan_type.id,
-        initial_head_count: selectedData.initial_head_count,
-        effectivity_date: selectedData.effectivity_date,
-        coc_issue_date: selectedData.coc_issue_date,
-        expiration_date: selectedData.expiration_date,
-        delivery_date_of_membership_ids:
-          selectedData.delivery_date_of_membership_ids,
-        orientation_date: selectedData.orientation_date,
-        initial_contract_value: selectedData.initial_contract_value,
-        mode_of_payment_id: (selectedData as any).mode_of_payment.id,
-        wellness_lecture_date: selectedData.wellness_lecture_date,
-        annual_physical_examination_date:
-          selectedData.annual_physical_examination_date,
-        commision_rate: selectedData.commision_rate,
-        additional_benefits: selectedData.additional_benefits,
-        special_benefits: selectedData.special_benefits,
-        name_of_signatory: selectedData.name_of_signatory,
-        designation_of_contact_person:
-          selectedData.designation_of_contact_person,
-        email_address_of_contact_person:
-          selectedData.email_address_of_contact_person,
-      },
-    ])
-
-    if (isInsertAccountError) throw new Error('Error updating pending account')
-
-    // Update pending account
-    await updatePendingAccount({
-      id: selectedData.id,
-      is_approved: true,
-    })
-  }
-
-  useEffect(() => {
-    if (isInsertingAccount || isUpdatingPendingAccount) setIsLoading(true)
-    else setIsLoading(false)
-  }, [isInsertingAccount, isUpdatingPendingAccount])
-
-  return <div onClick={handleClick}>{children}</div>
-}
-
-export default ApproveRequestButton
