+++ src/app/(dashboard)/admin/approval-request/accounts/modal/approval-request-info.tsx
@@ -4,7 +4,7 @@ import {
   formatPercentage,
 } from '@/app/(dashboard)/(home)/accounts/columns/accounts-columns'
 import { useApprovalRequestContext } from '@/app/(dashboard)/admin/approval-request/accounts/approval-request-provider'
-import ApproveRequestButton from '@/app/(dashboard)/admin/approval-request/accounts/approve-request-button'
+import ActionRequestButton from '@/app/(dashboard)/admin/approval-request/accounts/action-request-button'
 import ApprovalInformationItem from '@/app/(dashboard)/admin/approval-request/accounts/modal/approval-information-item'
 import { Button } from '@/components/ui/button'
 import {
@@ -219,12 +219,16 @@ const ApprovalRequestInfo = () => {
 
             {/* Actions */}
             <div className="flex justify-end gap-x-2">
-              <Button variant={'destructive'}>Reject</Button>
-              <ApproveRequestButton>
+              <ActionRequestButton action="reject">
+                <Button variant={'destructive'} disabled={isLoading}>
+                  {isLoading ? <Loader2 className="animate-spin" /> : 'Reject'}
+                </Button>
+              </ActionRequestButton>
+              <ActionRequestButton action="approve">
                 <Button variant={'default'} disabled={isLoading}>
                   {isLoading ? <Loader2 className="animate-spin" /> : 'Approve'}
                 </Button>
-              </ApproveRequestButton>
+              </ActionRequestButton>
             </div>
           </div>
         </div>
